coordinator: Switch tests to use non-transactional RPCs
The coordinator tests have two tests which verify routing logic for
RPCs which don't use transactions. This is currently implemented via
`FetchIntoObjectPool`, but we're in fact about to change it to
conditionally use transactions. As this would break tests, let's switch
to a different RPC whihc we know won't use transactions.