testhelper: Stop trying to create testdata directory
The `testhelper.EnableGitProtocolV2Support()` function replaces the Git
binary with a custom script in order to check whether certain envvars
are set. The variables were historically written into our `testdata`
directory, but it has since been converted to instead use the global
temporary test directory. One artifact which survived though is that we
try to create the `testdata` directory if it doesn't exist. So let's
remove that, it's not need anymore.