transactions: Allow disabling with an env var
Logic-wise this comments is intended to achieve the same as reverting:
ca8e2de58193775a62599b3a4de682c511046033:
(featureflag: Remove reference transaction feature flag), just without
conflicts and a more concentrated diff.

The use case is that a customer turned the feature flag for reference
transactions off, which wasn't honoured anymore. This change allows for
disabling transactions again.

The root problem lies in the fact that Gitaly needs to
connect with one Praefect that coordinates the current transaction.
Given this is an IP based connection, not to a domain name, this
violates some security policies.