Introduction of FailOnNonTempDialError to omit indefinite dial attempts on bad endpoints
When pool creates connection to the remote internally it doesn't verify
if connection is properly configured and can be used for the requests.
Created connection then added into the pool and will be returned on the next
attempt to get connection to the same address. To verify if connection
is valid and could be used after dialing a new set of options composed
into a FailOnNonTempDialError function and should be used with Dial call.
The Dial call will block until connection successfully initialized and
truly established with the remote. And to address indefinite blocking
of Dial if there is no remote listening on that address the FailOnNonTempDialError
option is used with 'true' flag.

The new set of options is not used in praefect connection manager to dial
with remote gitaly servers because we don't want stop praefect from start
in case one of gitaly nodes is down.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2948