UserCreateBranch: unify API responses between Go & Ruby response paths
Change code added in [1] to use GetReference() instead of GetBranch(),
and make it return the same error messages & codes as the Ruby code.

This resolves numerous bugs in [1] which weren't spotted due to
inadequate test coverage. Those are now fixed.

Other bugs in the code added in [2] remain, but I'm still working on
fixing those.

See also [3] for the initial discarded MR to fix issues in this code
for both UserCreate and UserDelete at the same time (that WIP itself
has bugs...).

Notes:

 A. The 08f22f2 and c642fe9 SHAs aren't special here. They just happen
    to be the next commits after 1e292f8 in the test repo[4]. I just
    needed some test commits.

 B. Converting away from helper.ErrPreconditionFailed(err) &
    helper.ErrPreconditionFailedf(fmt, err) here is intentional. See
    [5].

1. c1e3ccca (Initial go implementation of UserCreateBranch, 2020-09-30)
2. c3b32722 (Initial go implementation of UserDeleteBranch, 2020-09-29)
3. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2838
4. https://gitlab.com/gitlab-org/gitlab-test
5. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2857#note_457436954