No longer compare checksums after replication
Prior to this change, the checksums of primary and replica were
calculated after completing the replication. In some cases, i.e. when
the primary did receive more updates after replication was initiated,
these checksums did not match.

This change removes the checksum calculation and comparison because it
was only added for demos in the first place. Removing this checksumming
will reduce load on the cluster.

Closes https://gitlab.com/gitlab-org/gitaly/-/issues/3110