Add link to suggested reviewer MR dashboard
Prior to this change, it was hard to estimate how much MRs a person had
assigned for review. Also, there were no reviewers suggested if the MR
was assigned to yourself already.

This change adds a link to the suggested reviewer's MR dashboards. These
dashboard show the MRs assigned to them, but authored by someone
else. This should give an indication of how many reviews they have on
their plate already.

The reviewers suggestion will now also no longer disappear, and we use
the MR IID to seed the random number to suggest reviewers, which should
give the same names every time.