operations: Fix UserDeleteBranch's use of hooks
UserDeleteBranch updates references via `updateReferenceWithHooks`, but
passes incorrect parameters to that function. Instead of giving OIDs for
the old, it instead passes the actual branch name. As this value is
later piped to our git hook implementations which expect the same format
as documented in githooks(5), this is wrong.

The issue could've only existed because our hooks don't yet properly
validate their input parameters, but this is about to change. So let's
fix the abuse as tighter input validation will cause this RPC to fail.