UserDeleteBranch: unify API responses between Go & Ruby response paths
Continue the changes I started in [1] and fix discrepancies in the
Ruby and Go codepaths for UserDeleteBranch. This resolves bugs in [2]
that weren't spotted due to missing test coverage, which is also being
fixed here.

We're now faithfully (sigh!) emulating the weird emergent behavior in
the Ruby code that it's impossible to delete a
refs/heads/refs/heads/FOO branch that you've created, since it'll be
auto-translated to refs/heads/FOO. See [3] for an explanation of the
Ruby code involved.

The rest of the change is amending & testing for existing error
messages we generate. This was more unified than UserCreateBranch
already since the Go code has been intercepting empty branch names,
empty users, whereas in UserCreateBranch it's handled separately by
the Ruby & Go code paths.

I'm changing away from helper.ErrPreconditionFailed() here while I'm
at it because it and its "*f()" cousin are not just wrappers for
"status.Errorf(codes.FailedPrecondition, ...)". Rather they're meant
for re-throwing nested errors, here we're not doing any rethrowing of
errors further down the stack. So let's not use the API intended for
that purpose. See also [4].

1. 718d06898 (UserCreateBranch: unify API responses between Go & Ruby
   response paths, 2020-11-30)

2. c3b32722 (Initial go implementation of UserDeleteBranch,
   2020-09-29)

3. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2864#note_459156912

4. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2859