hook: Simplify setup of envvars for pre-receive hooks
The current way we set up environment variables for the pre-receive hook
is a bit convoluted and much more verbose than it needs to be. While it
makes sense to have a shared function which sets up envvars for both
pre- and post-receive hooks via `hookRequestEnv()`, the additional
`prePostRequest` interface is only used in a single place.

Let's just inline it, get rid of that interface and append push options
directly in order to simplify the code.