git: Stricter validation for expected envvars
Our testcase for the reference-transaction hook currently verifies that
we get a subset of environment variables. While it's not sensible to
test for all environment variables as they will also contain
user-specific variables like PATH, TZ and so on, we can at least do
better with regards to Gitaly-specific variables. So let's improve the
test and make it stricter by verifying we have an exact match for all
variables starting with "GITALY_" or "GL_".