run sql electors checks in a transaction
The SQL elector runs multiple statements in separate transactions.
This causes problems when the election is happening in a slow manner,
perhaps due to a slow database.

The successful health check timestamps are updated first in different
transactions. These are then checked later in other transcations to
determine the healthy nodes. If Praefect takes longer to perform the
check than what is considered to be the healthy node timeout, Praefect
will see the nodes as being unhealthy and demote the primary. Primary is
then possibly re-elected immediately afterwards as the ticker loop in
sqlElector.monitor does not compensate for slow execution but queues up another
execution to happen immediately after the slow one.
This causes the primary node to be demoted and re-elected when the election
process is not fast enough.

This commit fixes the problem by running each of the health check and failover
runs in a single transaction. This ensure the current time is the same for each
statement and there should be no re-election loop when a single Praefect node
is operating slowly.