Internal connections to itself should use token
When we use gRPC to connect gitaly to itself it still
should use a secret token to dial that connection, no matter
if it connects throw socket or not. The token will be
verified on each RPC call done to the server. If not set
the RPC will result into an error.