Remove passing internal socket location as a separate param
A couple of gitaly services require to connect to itself (same
gitaly instance) to execute other RPC calls. For doing this
it uses a unix socket. It is configured automatically on
server startup and listens on incoming requests as all other
endpoints (tcp, tls). As the socket location is part of the
config.Cfg setup we don't need to pass it as a separate parameter
where we pass config.Cfg variable as it could be taken from it.