hooks: Setup and use payload
While the preceding commit implemented the new HooksPayload structure
which is going to contain all Gitaly-relevant parameters eventually, it
didn't yet set up the environments. This commit here thus starts
injecting the payload into the hooks' environment.

As the payload has a fallback to use the now-deprecated environment
variables GITALY_REPO and the likes, we can also go all-in and directly
start to use theh payload returned by `HooksPayloadFromEnv()`. If we
don't have a payload injected but got those old envvars, it would work
fine regardless.