gitaly-hooks: Improve test setup to match real setups more closely
The test setup for gitaly-hooks doesn't match at all what "real" setups
look like. Neither does it setup the authentication token as part of the
Gitaly configuration nor does it use the internal socket path for the
hook service. While it doesn't matter right now, it will cause problems
in a future commit when we're starting to migrate to a unified structure
containing all hook-relevant information.

Let's refactor the test setup as a preparatory step to be prepared for
that future.