repository: Remove test relying on broken behaviour
In 750b5b44c (Internal connections to itself should use token,
2020-12-02), we fixed a bug which caused us to not pass the
authentication token when connecting to the internal socket. While the
fix is perfectly sensible, unfortunately the test that was added relies
on broken behaviour.

When executing `CloneFromPoolInternal`, it will call out to
`FetchInternalRemote` which in turn executes git-fetch(1) to perform the
fetch and some other RPC calls to get information about the remote's
state. To do this, we have two tokens at play: first the credentials of
ourselves which we need to connect to our own internal socket. And
second the credentials of the remote repository. In case of the test,
remote and internal socket are in fact the same. This info is accessed
once via the config and once via context-injected Gitaly server
information.

In the test setup, we set up the server with invalid authentication
token via the config, but inject other information into the context.
The expectation now is that the test will fail with the RPCs failing.
Which is weird, as in fact we should already fail with the git-fetch(1)
as it uses gitaly-ssh as transport and thus connects to the internal
socket with invalid credentials. But it happens to work right now
because of another bug: git-fetch(1) uses both a custom environment and
`git.WithRefTxHook()`, where the latter will override the custom
environment. Thus it accidentally uses correct credentials.

This test is thus going to stop working as soon as we fix this bug. As I
wasn't able to come up with a test setup which demonstrates the same
failure, this commit just removes the test.