Verify connections are properly configured with HealthCheckDialer
There is no way we could identify if the connection we establish
to remote is properly configured and could be used with no issues.
As a workaround on it is to issue a request to the remote when
connection is set. The simplest approach is to issue a health check
request as both gitaly and praefect expose this service outside.
As we don't want to make it after each call to Dial as it could be
missed or if we change implementation later it makes sense to define
a custom Dialer for that purpose. It should be used with the pools
we are using except of praefect connection pool to gitaly nodes,
as it would prevent praefect from start if one of gitaly nodes is not
accessible.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2699