Connections pool should respect token as part of the key
It could be possible the same address will require a new token
to be used. This could be possible in case of the new token
rollout. The simplest approach is to make a token part of the
key in the connections pool. With that the same address will
be able to have multiple connections with different token.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2699