style: use a better alias for "git reference"
Since v2.25.0 of git you've been able to use this way of doing it. See
git-vcs/git@1f0fc1db85 (pretty: implement 'reference' format,
2019-11-19).

So let's use that example, and while I'm at it change the SHA-1 in the
example here to the one that landed in the repo, not some pre-rebase
version that was only in @pks-t's local repo.