coordinator: Conditionally enable transactions for RPCs
With transaction information now being available for all of our Git
commands which may modify references, we can now finally enable
transactions for all RPCs which may modify references. In order to ease
the transaction and be able to iterate fast, this commit thus implements
feature flags for each of the RPCs which do.