operations: Remove GoUserMergeToRef feature flag
The Go port of UserMergeToRef has been introduced in 13.5, enabled by
default in 13.6 and until now didn't cause any known bugs or
regressions. So let's remove the feature flag altogether.

Note that we still cannot completely get rid of the Ruby codepath. After
porting the feature to Go, the Ruby codepath grew another feature which
is to accept conflicts. As Go doesn't handle this yet, we thus need to
retain the Ruby codepath for now.