Refactor alternates package to deprecate helper.GetRepoPath
The alternates package relies on helper.GetRepoPath to define
path of the repository and actually does two things at a time:
returns a path to repo and set of env vars that point to alternate
object directories.
The new function Env added that only returns a set of env vars
and accept all dependencies as parameters instead of calling
functions dependent on global state.

The PathAndEnv function is not removed because it is used by the
catfile package and requires a big refactoring as well. For now
it is marked as deprecated.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699