UserCreateTag tests: test for tags to trees & blobs
The GitLab UI supports manually entering a tag pointing to a tree or a
blob, but in crafting the response Gitaly's Ruby code will error after
creating the tag.

Regardless of that the tag *is* created, so let's test for that while
avoiding the question of what the response is. A follow-up commit will
fix the error in the Ruby and try to do something more sensible there.

I did not test the very limited response here because even if I could
get it to dump out enough to see that e.g. this is true:

    require.False(t, response.Exists)
    require.Equal(t, response.PreReceiveError, "")
    require.Equal(t, response.Tag.Id, "")

Inspecting it would lead to:

    panic: runtime error: invalid memory address or nil pointer dereference [recovered]
            panic: runtime error: invalid memory address or nil pointer dereference
    [signal SIGSEGV: segmentation violation code=0x1 addr=0x0 pc=0xbbc9bb]

I did not investigate, presumably we're being overly trustful of the
Ruby process's response, which just died or something. Anyway, it's
being fixed in the next commit. So let's move on.