UserCreateTag: fix internal error in tree/blob tag creation
As noted in the last commit & the comment being added here the Ruby
UserCreateTag code would successfully create tags pointing to blobs
and trees, but throw an error while doing so.

The response we're returning is still rather nonsensical in this case,
and because of the limitations of the protocol we can't return a good
response here. But let's fix things up enough so that we're not just
returning this to the client:

    rpc error: code = Unknown desc = NoMethodError: undefined method `rugged_commit' for nil:NilClass

Instead we fake up a response as best we can. There's some
intermediate marshaling between the code I'm changing & what we get
which even inserts the right SHA-1s for us.