hooks: Use common code path to configure hook setup
While both `WithReceivePackHooks` and `WithRefTxHooks` call
`refHookEnv`, they still need to manually handle appending the generated
environment to the actual command. We can do better here and just
directly pass in the command and have it configure it, which also allows
us to deduplicate both generation of the transaction environment and
setup of "core.hookPath". This allows us to put also get rid of
generating the hooks path in `ReceivePackConfig()` as we know that all
of its users also use hooks.