hooks: Make `HooksPayloadFromEnv()` more extendable
We're going to add more variables to the HooksPayload structure, which
means that we'll also have to add more fallbacks to how they were
retrieved previously. As a preparatory step, this commit reworks
`HooksPayloadFromEnv()` such that it's easier to add in those additional
fallbacks.

No change in behaviour is expected from this change.