hooks: Fix incomplete test setups
When invoking hooks, the code assumes several environment variables to
exist. Most importantly, we need to provide information about where to
find the gitaly-hooks binary, but also meta information like the user
who caused the hook execution and the repository it was triggered in.
While many tests already set this up correctly, others don't. This
worked fine until now, but this is mostly by luck only. And as soon as
we're going to expect the HooksPayload to always exist in the
environment, those tests will fail.

Let's prepare for this change by adjusting our tests to always inject a
HooksPayload.