gitaly-hooks: Use common function to compute hook environments
We're currently the same snippets of code to assemble environment
variables across all of our hook implementations, which is to assemble
the list of allowed, "GL_" and transaction-related environment
variables. This code is repetitive, and most importantly we also forgot
to inject transaction envvars for the update hook. As this information
is required by the update hook to determine whether it is currently the
primary transaction node and thus whether it's expected to execute
custom hooks or not, it's crucial to have this info available.

Let's fix the issue by using a common function which assembles all of
these environment variables.