hooks: Start using hooks payload for transactions
Now that we have transaction-related information as part of the
HooksPayload, this commit converts code to use it instead of the
previous environment variables. This also requires us to inject the
variable into the hook requests, which wasn't yet the case.

While no code site yet injects transaction information via the payload,
this in fact serves as a nice demonstration that the fallback in
`HooksPayloadFromEnv()` works as expected.