Revert "featureflag: Remove reference transaction feature flag"
This reverts commit ca8e2de58193775a62599b3a4de682c511046033 and
restores the feature flag for enabling/disabling reference transactions.
The main reason to revert the change is that there are currently still
some problems with TLS setups and injected Praefect server information,
where Gitaly may not be able to verify authenticity of Praefect servers
and thus fail to establish a connection.

This also removes the environment variable to disable transactions,
which was only added as a short-term fix to make GitLab setups work in
such environments again.