housekeeping: Do not delete temporary files outside of ODB
When git processes crash or the servers gets rebooted without an orderly
shutdown, then it may happen that git processes may leave behind stale
files. In the context of object files, packfiles and pack indices, these
are "tmp_" files in the ".git/objects" folder which may never get
cleaned up. This is why we have the housekeeping tasks, to clean up any
of these stale temporary files which are older than a certain threshold.

As it turns out though, our pruning logic is quite happy to remove files
other than temporary objects as it doesn't care about where it actually
finds a given "tmp_" file. So if you have e.g. a reference which is
called "tmp_branch", then we may eventually remove this branch if it
only exists as a loose reference. Right now we only perform housekeeping
as part of garbage collection calls after calling git-gc(1). But it's
not guaranteed that it'll actually pack loose references into packed
ones, so we're building on quite weak foundations here.

Fix this potential for data loss by only searching for temporary objects
inside of the object database.