objectpool: Perform housekeeping
When performing garbage collection on repositories, we always call our
housekeeping tasks to rid the repository of any files which we know
shouldn't be there anymore. This ensures we remove stale files which
only take up precious disk space, but it also in some cases ensures that
we're removing files which may cause git to malfunction. One such case
is for example empty references which may be created under some
circumstances during hard reboots or crashes.

While we do this for repositories, we don't do so for our object pools.
This is causing hard to spot issues with repositories not getting
updated or not getting repacked. After a while, this leads to a
deterioration of performance when cloning repositories which are part of
such a pool.

There isn't any place where we currently perform garbage collection, but
the next-best place is `FetchFromOrigin()`. Next to pulling in new
objects from the source repository, it also packs references and checks
for repository consistency. So let's use it to also perform our
housekeeping tasks for object pools here.