conflicts: Don't set "uploadpack.allowAnySHA1InWant" in environment
In the conflicts service, we're currently setting up the environment
variable "uploadpack.allowAnySHA1InWant=true". This is wrong in multiple
ways: first, setting this variable on the client-side doesn't make any
sense as this is purely a server-side option for git-upload-pack(1). And
second, setting git config options as part of the environment doesn't
have any effect at all.

Fix the issue by removing the setup altogether and inlining the constant
at the only site where its used. This also removes the confusion by
having the constant be called `EnvVar...`, which probably is why it's
been set up as part of the environment in the first place.