git: Drop `CmdStream` parameter from `SafeCmdWithoutRepo`
Before we had variable command options, we refactored some of our
functions to accept a `CmdStream` parameter. This is confusing nowaday
as there is now multiple ways to pass standard streams to commands, once
via the parameter and once via `WithStdout()` and the likes.

Let's drop the parameter in favor of using options to avoid the
confusion and to unify interfaces.