git: Use interface-specific functions to build arguments
In order to disambiguate implementations of the Option and Cmd
interfaces, we currently use both a `IsCmd()` and an `IsOption()`
function which simply does nothing. This is bad style, which is why this
commit instead refactors the code to have type-specific functions
`CommandArgs()` and `OptionArgs()`. This is more idiomatic and also
makes code easier to read when arguments are assembled.