git: Pull out function to assemble command arguments
Assembling arguments for subcommands is non-trivial logic as it requires
handling names, options, positional arguments and post-separator
arguments. As we're about to introduce a second user which requires this
logic by converting `SubSubCmd` from an option to a command, let's pull
out common code into a separate function which can be called by both
SubCmd and SubSubCmd types.