Deprecate helper.GetStorageByName
Because helper.GetStorageByName depends on the global var
config.Config we should remove it from the code in order to
clean up config.Config usage. Currently it is not possible to
fully remove it from the code that is why helper.GetStorageByName
is marked as deprecated in favor of storage.Locator.
All possible usages are replaced with usage of storage.Locator.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699