Cleanup redundant data from notification events
Postgres sends notification messages to the Praefect using triggers
on 'repositories' and 'storage_repositories' tables changes.
The payload is constructed from the data that is subject of the change:
old and new states. The notifications are used to invalidate in-memory
cached values of up to date storages used by the reads distribution. Most
of the data from the notification payload is redundant and only two fields
are used: relative_path and virtual_storage.

This change cleans up the payload by including only required fields.
It also changes the structure of the notification message to reduce
redundancy as the payload of the notification message is limited by the
Postgres instance and big payloads could generate an error 'pq: payload
string too long'.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3309