docs: clarify "plumbing" v.s. "porcelain" in git(1) usage
Clarify the recently added documentation in 2cbfe50d7 (docs: Explain
how to interact with Git in Gitaly, 2020-09-23) to not draw a hard
line on using non-plumbing. As the change being made here argues this
doesn't reflect reality on the ground.

As a practical matter we make plenty of use of non-plumbing,
e.g. "gc", "bundle", "log", "remote", "config", "worktree" etc. What
we actually care about is the sort of command use I altered in
61b9b30a5 (git.Remote: use "git remote" exit code instead of parsing
stderr, 2020-11-10). Let's try to make the docs reflect that.

I also sent a note to the git ML about the general issue:
https://public-inbox.org/git/878sa7xujm.fsf@evledraar.gmail.com/T/