housekeeping: Don't try to fix permissions when removing files
When removing directory hierarchies, we may be unable to do so if any of
the directories in that hierarchy has permissions which don't allow us
to modify it. This used to be a problem in our housekeeping code as we
also triedi to remove temporary directories. But with the preceding
commit, we've removed this code as git shouldn't create any temporary
directories anyway.

Let's drop the force-removal code altogether. Removing files with
insufficient permissions works just fine as the only thing that matters
in this context is the permission of the containing directory.