housekeeping: Prune empty references
While git is quite careful with how it writes objects as it always uses
a temporary file first and only renames it into place as soon as all
data has been written to disk, it's not as careful when writing
references. While it does lock references before writing them, the
result is that under some circumstances we may end up with references
which have proper names, but don't have any contents at all. The exact
cause is unclear, but it's quite plausible that crashes, hard reboots or
similar things may cause this to happen.

Let's extend our housekeeping to also check for such empty references
and delete them. Detecting them is quite easy: they have a size of zero
bytes and thus cannot contain anything meaningful. There still is the
possibility of a race with any concurrently running git process though,
which is why the new cleanup logic will only cleanup empty references
older than a day. That should be more than enough time to create the
file and fill it with data.