Fix tests for cgroups that depends on linux-only library
The implementation of the cgroups for Gitaly uses third
party dependency https://github.com/containerd/cgroups
that could be used only on the linux systems.
The production code compiles without issues as library
dependent code protected with '_linux' suffix that removes
it from compilation on MacOS systems, but the test code
has no such suffixes that is why it breaks tests on OSX
system because of the library packages use in the tests.
This commit removes those files from the scope on OSX
systems by using the same strategy with file name suffix.