hooks: Disallow configuring hooks multiple times
Now that we have a shared hook setup which always sets up the hooks
path, transaction environment as well as the mandatory hooks payload, we
should enforce that hooks aren't being setup multiple times. While it
doesn't hurt right now as the subset of information that's shared across
both hook configurations is exactly the same, this will change in the
future as we put more information into the hooks payload. And as soon as
we do, configuring multiple hooks may easily override information based
on the order in which hooks were configured.

So let'just assert that hooks are always configured exactly once for
each command which may update references.