hooks: Remove GITLAB_SHELL_DIR from environment
Back when hooks were still implemented via gitlab-shell, the
GITLAB_SHELL_DIR environment variable was used to tell the hooks where
to find the configuration file as well as the gitlab-shell secret. This
was used for example to call out to the GitLab API.

Nowadays we don't use gitlab-shell for this anymore, but instead only
use our Go hooks. As such, it's not necessary anymore to pass along the
configuration, so we can now remove both the environment variable as
well as supporting code.

There is one outlier in this patch, which is the GITLAB_PROJECT_TIMEOUT
configuration, which is used to timeout repository mirroring. It's not
anything inherent to the gitlab-shell and in fact it's only used in a
single place. So let's just inline it and be done -- it's not currently
configurable anyway.