Fix usage of MockRepositoryStore in a test setup
When RepositoryStore configured with the optional parameter
it is re-written internally with another instance of the
MockRepositoryStore, so you have two different implementations
of the same abstraction in a single test setup that leads to
weird behavior of the components. This commit removes fixes
this logic by removing instance that overrides passed from outside
or default RepositoryStore implementation.