Tags tests: don't assign back to testCase reference in test loop
Fix a logic error of mine in b1f09d511 (UserCreateTag: fix internal
error in tree/blob tag creation, 2020-12-04).

It makes no sense to be assigning to our test data itself during the
loop. As @toon points out in [1] the testCase variable is a reference.

This didn't introduce a logic error in the test itself, since we're
only using each item in the testCase array once, so it's OK for the
test results if we mutate it as we go along. But let's change this to
something more sensible anyway & change our transitory responseOk
variable instead.

This anti-pattern was itself copied from earlier code in
2ad030438 (Implement UserCreateTag RPC, 2017-09-26). A follow-up
commit will fix that test case.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2881#note_463363831